Se crea el crud para los productos en donde se le asignan a los clientes solo pueden listar los productos

+++ src/main/java/com/ferreteria/demo/Repositories/Entities/Barrio.java
@@ -18,7 +18,7 @@ public class Barrio {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer idBarrio;
+    private Long idBarrio;
 
     @Column(nullable = false)
     private String nombre;
